refactor: swagger api 패키지로 변경
